Subclass Adjustment
Created PickupArea2D to extend into further items that can be picked up. Just make a new script and target the ApplyyPickupEffect method to do whatever you need to the player.
